+++ crevice-derive/src/lib.rs
@@ -1,7 +1,8 @@
 use proc_macro::TokenStream as CompilerTokenStream;
 
+use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
-use syn::{parse_macro_input, Data, DeriveInput, Fields};
+use syn::{parse_macro_input, Data, DeriveInput, Fields, FieldsNamed};
 
 #[proc_macro_derive(AsStd140)]
 pub fn derive_as_std140(input: CompilerTokenStream) -> CompilerTokenStream {
@@ -75,6 +76,8 @@ pub fn derive_as_std140(input: CompilerTokenStream) -> CompilerTokenStream {
         initializer.push(quote!(#field_name: self.#field_name.as_std140()));
     }
 
+    let struct_alignment = emit_struct_alignment(16, fields);
+
     let type_layout_derive = if cfg!(feature = "test_type_layout") {
         quote!(#[derive(::type_layout::TypeLayout)])
     } else {
@@ -101,7 +104,7 @@ pub fn derive_as_std140(input: CompilerTokenStream) -> CompilerTokenStream {
         unsafe impl #impl_generics ::crevice::internal::bytemuck::Pod for #std140_name #ty_generics #where_clause {}
 
         unsafe impl #impl_generics ::crevice::std140::Std140 for #std140_name #ty_generics #where_clause {
-            const ALIGNMENT: usize = 16;
+            const ALIGNMENT: usize = #struct_alignment;
         }
 
         impl #impl_generics ::crevice::std140::AsStd140 for #name #ty_generics #where_clause {
@@ -119,4 +122,30 @@ pub fn derive_as_std140(input: CompilerTokenStream) -> CompilerTokenStream {
 
     // Hand the output tokens back to the compiler
     CompilerTokenStream::from(expanded)
+}
+
+/// Emit a const expression that computes the alignment of a struct based on its
+/// minimum alignment and fields.
+///
+/// The minimum alignment of an std140 struct is 16, while the minimum alignment
+/// for an std430 struct is 0.
+fn emit_struct_alignment(min_alignment: usize, fields: &FieldsNamed) -> TokenStream {
+    // This fold builds up an expression that finds the maximum alignment out of
+    // all of the fields in the struct. For this struct:
+    //
+    // struct Foo { a: ty1, b: ty2 }
+    //
+    // ...we should generate an expression like this:
+    //
+    // max(ty2_align, max(ty1_align, min_align))
+    fields.named.iter().fold(quote!(#min_alignment), |last, field| {
+        let field_ty = &field.ty;
+
+        quote! {
+            ::crevice::internal::max(
+                <<#field_ty as ::crevice::std140::AsStd140>::Std140Type as ::crevice::std140::Std140>::ALIGNMENT,
+                #last,
+            )
+        }
+    })
 }